// src/system/model/items/itemType.ts
export enum ItemType {
  armor = "armor",
  implant = "implant",
  firearm = "firearm",
  software = "software",
  vehicle = "vehicle",
  weapon = "weapon",
  device = "device",
  license = "license",
  firearmAccessory = "firearmAccessory",
  sin = "sin",
  credstick = "credstick",
  program = "program",
  other = "other",
}