// src/system/model/items/weaponData.ts
import type { AttributeKey } from "#/system/model/attributes/attributeKey.ts"
import type { SkillKey } from "#/system/model/skills/skillKey.ts"

import type { ItemData } from "./itemData.ts"
import { ItemType } from "./itemType.ts"
import type { FirearmTypeKey } from "./weapons/firearms/firearmTypeKey.ts"

export enum FirearmAttachmentPoint {
  Internal = "Internal",
  Top = "Top",
  Under = "Under",
  Barrel = "Barrel",
}

export enum WeaponType {
  melee = "melee",
  thrown = "thrown",
  projectile = "projectile",
  firearm = "firearm",
  exotic = "exotic",
  other = "other",
}

/** The sub-type of a melee weapon, which determines the associated combat skill. */
export enum MeleeWeaponType {
  blade = "blade",
  club = "club",
}

export interface WeaponData extends ItemData {
  dmg: string
  dmgType?: "physical" | "stun" | "custom"
  ap?: number
  itemType: ItemType.weapon
  weaponType: WeaponType
  skill: SkillKey
  attribute?: AttributeKey
}

export interface FirearmData extends WeaponData {
  weaponType: WeaponType.firearm
  firearmType: FirearmTypeKey
  firemodes: string[]
  recoil: number
  attachmentPoints?: FirearmAttachmentPoint[]

  ammo: {
    size: number
    remaining: number
    type:
      | "break"
      | "clip"
      | "drum"
      | "muzzle"
      | "magazine"
      | "cylinder"
      | "belt"
  }
}

export interface MeleeWeaponData extends WeaponData {
  weaponType: WeaponType.melee
  reach: number
  meleeType?: MeleeWeaponType
}

export interface FirearmAccessoryData extends ItemData {
  itemType: ItemType.firearmAccessory
  enabled?: boolean
  mountPoints: FirearmAttachmentPoint[]
  parentSlot?: FirearmAttachmentPoint
}

export type AnyWeaponData =
  | FirearmData
  | MeleeWeaponData

export type WeaponDataFor<T extends WeaponType> = Extract<AnyWeaponData, { weaponType: T }>

export function isWeaponData(item: ItemData): item is WeaponData {
  return item.itemType === ItemType.weapon
}

export function isFirearmData(item: ItemData): item is FirearmData {
  return isWeaponData(item) && item.weaponType === WeaponType.firearm
}

export function isMeleeWeaponData(item: ItemData): item is MeleeWeaponData {
  return isWeaponData(item) && item.weaponType === WeaponType.melee
}